+++ packages/core/changelog.md
@@ -407,3 +407,15 @@ List React19 as Dependency.
 ### Core
 
 - fix(useEventSource): improve EventSource connection handling
+
+## 6.0.4(Jul 02, 2025)
+
+### Core
+
+- fix(useClipboard): immediately access clipboard
+
+## 6.0.5(Jul 03, 2025)
+
+### Core
+
+- fix(useClipboard): add document focus check